Refactor unit tests
