+++ utils/__tests__/formatPercent.test.ts
